requirements.txt
Pillow
pyautogui
opencv-python
yt-dlp
